requirements.txt
comfyui
timm
sentencepiece>=0.1.97
transformers>=4.34.1
accelerate>=0.23.0
